python/pygame code to test usb webcam
Works on Raspberry Pi with no additional software requirements. 
Simple run 
python pycam.py
It should startup the camera and show the video.
To exit either click the [x] if in the desktop or press any key.
You can chance the width and height of the capture and window in the code.

Note: I was using this to test a cheap £1.84 (At the time of purchase) VGA USB webcam http://ebay.to/2aYCM6V
The webcam works in Scratch and with motion on the Pi.